+++ changelog.md
@@ -8,6 +8,7 @@
 * (Developer) Implemented new HtmlHelper extension `AddCustomHeadParts`: registers whatever head (meta) tag you wish
 * (Developer) Added `SmartUrlRoutingModule`, which can pass static files to `UrlRoutingModule` if desired (e.g. used by MiniProfiler). This way static files can be handled by regular actions or filters, without polluting web.config.
 * New payment plugin "Payone"
+* Option to set a delivery time for products available for order with stock quantity < 1
 
 ### Improvements
 * (Perf) Faster application warmup